refactor: Make getNextIndex shareable internally
