+++ lib/components/private/getNextIndex.ts
